Somewhat hard-code special case for fedora (#47)
For some reason, on some unixes and in some 'trash' folders the
original file paths include the user name and the trash directory
itself is buried in .local/share. This violates the previous assumption
that the trash folder is in the users home and leads to invalid
original paths, causing restore to fail.

Now that I am thinking about it… isn't it use who write the trash info
file?