fix: Properly reconstruct paths when restoring files on freedesktop if those were relative (#47)
Previously it would be unable to reconstruct original paths if the trash
directory was on a mount point due to a 'split brain' of sorts.

When trashing files it would create original path information based
on them being relative to a mount point, but when restoring them
it would reconstruct them to be relative to the trash top level
directory.

Now the reconstruction happens against to mount point itself which makes
restoration match.