fix: avoid inconsistency when using relative paths in trashed file info. (#39)
We use absolute paths now without trying to generate a relative path
based on some top directory as the latter seems to be causing
inconsistencies on some linux distros, as the restore path ends
up being incorrect.

Rather go with the absolute truth and don't fiddle with path
transformations at all to make it work everywhere.