fix: Support non-Unicode paths
There are several spots where paths are assumed to be Unicode. However,
some (all?) operating systems support non-Unicode paths which causes
`trash-rs` to panic if encountered. I switched some of those code to use
`OsString`s instead of `String`s. Unfortunately, I had to add a new
dependency, `urlencoding`, in order to properly handle decoding non-UTF8
byte slices.

As of this commit, the test suite passes and code should be ready, but I
will try to remove the `url` crate and use `urlencoding` in its place
in the next commit.

Closes: #105