Update Windows code to account for API change
I opted to not remove the UTF8 verification for Windows' platform
specific code. While it's unlikely that Windows' API would return
invalid Strings, the extra check for a filename can't hurt whereas
removing it would require modifying a decent chunk of the code.

The old code performed the check, and converting a String to an OsString
is free. Path of least resistance.