Accelerate by removing recursive renaming
`execute_src_to_dst_operation` does the operation recursively for every file in
a directory which makes `trash`ing a large item with hundreds of thousands of
files in it as slow as `rm`, especially on hard disks. Actually it is not
necessary since `try_creating_placeholders` has ensured thread safty -- only the
one who has successfully created the placeholder has the chance to run the
following code.