feat: Short circuiting check for empty trash
`is_empty()` is a short circuiting function that checks if the trash is
empty on Freedesktop compatible systems and Windows.

The main purpose of `is_empty()` is to avoid evaluating the entire trash
context when the caller is only interested in whether the trash is empty
or not. This is especially useful for full trashes with many items.